+++ docs/changes.rst
@@ -4,6 +4,8 @@ Changelog
 1.8.1
 =====
 * Drop `pkg_resources` and use `importlib.metadata` to access package version string.
+* Add missing `region_name` in `s3fs` store creation to set required location contraint
+  during bucket creation.
 
 1.8.0
 =====